mise en page commune des jsp modif et inscritpion + css

+++ src/main/java/fr/eni/eniD2WM147/bll/EnchereManager.java
@@ -49,7 +49,24 @@ public class EnchereManager {
 		return user;
 	}
 
-	public void deleteUser(int idUtilisateur) {
+	public Utilisateur updateUserProfil(String pseudo, String nom, String prenom, String email, String tel, String rue,
+			String codePostal, String ville, String mdp, int credit, int idUtilisateur) throws BusinessException {
+		BusinessException bE = new BusinessException();
+		if (!bE.getListeMessage().isEmpty()) {
+			throw bE;
+		}
+		
+		Utilisateur user = utilisateurDAO.updateUserProfil(pseudo, nom, prenom, email, tel, rue, codePostal, ville, mdp, credit, idUtilisateur);
+				
+		return user;
+
+	}
+
+	public void deleteUser(int idUtilisateur) throws BusinessException {
+		BusinessException bE = new BusinessException();
+		if (!bE.getListeMessage().isEmpty()) {
+			throw bE;
+		}
 		utilisateurDAO.deleteUser(idUtilisateur);
 	}
 
@@ -63,16 +80,20 @@ public class EnchereManager {
 		return articleDAO.selectAllArticles();
 	}
 
- 	public List<ArticleVendu> selectArticlesByCat(int noCategorie) throws BusinessException {
+	public List<ArticleVendu> selectArticlesByCat(int noCategorie) throws BusinessException {
+		BusinessException bE = new BusinessException();
+		if (!bE.getListeMessage().isEmpty()) {
+			throw bE;
+		}
 		return articleDAO.selectArticlesByCat(noCategorie);
 	}
- 	
- 	public List<Categorie> selectAllCat() throws BusinessException{
- 		BusinessException bE = new BusinessException();
+
+	public List<Categorie> selectAllCat() throws BusinessException {
+		BusinessException bE = new BusinessException();
 		if (!bE.getListeMessage().isEmpty()) {
 			throw bE;
 		}
 		return articleDAO.selectAllCategories();
- 	}
+	}
 
 }
